// src/types/user.ts
export interface User {
	title: "Mr" | "Mrs";
	firstName: string;
	lastName: string;
	email: string;
	password: string;
	birthDay: string;
	birthMonth: string;
	birthYear: string;
	company: string;
	address: string;
	address2?: string;
	country: string;
	state: string;
	city: string;
	zipcode: string;
	mobile: string;
}